+++ familyhelper-webapi-impl/src/main/java/com/dwarfeng/familyhelper/webapi/impl/service/assets/ItemResponseServiceImpl.java
@@ -22,6 +22,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 
@@ -35,13 +36,15 @@ public class ItemResponseServiceImpl implements ItemResponseService {
 
     private final AssetCatalogResponseService assetCatalogResponseService;
 
-    @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
     public ItemResponseServiceImpl(
-            @Qualifier("familyhelperAssetsItemMaintainService") ItemMaintainService itemMaintainService,
-            @Qualifier("familyhelperAssetsItemOperateService") ItemOperateService itemOperateService,
+            @Qualifier("familyhelperAssetsItemMaintainService")
+            ItemMaintainService itemMaintainService,
+            @Qualifier("familyhelperAssetsItemOperateService")
+            ItemOperateService itemOperateService,
             @Qualifier("familyhelperAssetsItemTypeIndicatorMaintainService")
-                    ItemTypeIndicatorMaintainService itemTypeIndicatorMaintainService,
-            @Qualifier("familyhelperAssetsItemLabelMaintainService") ItemLabelMaintainService itemLabelMaintainService,
+            ItemTypeIndicatorMaintainService itemTypeIndicatorMaintainService,
+            @Qualifier("familyhelperAssetsItemLabelMaintainService")
+            ItemLabelMaintainService itemLabelMaintainService,
             AssetCatalogResponseService assetCatalogResponseService
     ) {
         this.itemMaintainService = itemMaintainService;
@@ -89,76 +92,59 @@ public class ItemResponseServiceImpl implements ItemResponseService {
     }
 
     @Override
-    public DispItem getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException {
-        Item item = itemMaintainService.get(key);
-        return dispItemFromItem(item, inspectAccountKey);
-    }
-
-    @SuppressWarnings("DuplicatedCode")
-    @Override
-    public PagedData<DispItem> allDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException {
-        PagedData<Item> lookup = itemMaintainService.lookup(pagingInfo);
-        List<DispItem> dispItems = new ArrayList<>();
-        for (Item item : lookup.getData()) {
-            dispItems.add(dispItemFromItem(item, accountKey));
-        }
-        return new PagedData<>(
-                lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(), dispItems
+    public PagedData<Item> childForAssetCatalogNameLike(LongIdKey assetCatalogKey, String pattern, PagingInfo pagingInfo) throws ServiceException {
+        return itemMaintainService.lookup(
+                ItemMaintainService.CHILD_FOR_ASSET_CATALOG_NAME_LIKE,
+                new Object[]{assetCatalogKey, pattern},
+                pagingInfo
         );
     }
 
-    @SuppressWarnings("DuplicatedCode")
+    @Override
+    public DispItem getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException {
+        Item item = itemMaintainService.get(key);
+        return toDisp(item, inspectAccountKey);
+    }
+
+    @Override
+    public PagedData<DispItem> allDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException {
+        PagedData<Item> lookup = all(pagingInfo);
+        return toDispPagedData(lookup, accountKey);
+    }
+
     @Override
     public PagedData<DispItem> childForAssetCatalogDisp(
             StringIdKey accountKey, LongIdKey assetCatalogKey, PagingInfo pagingInfo
     ) throws ServiceException {
-        PagedData<Item> lookup = itemMaintainService.lookup(
-                ItemMaintainService.CHILD_FOR_ASSET_CATALOG, new Object[]{assetCatalogKey}, pagingInfo
-        );
-        List<DispItem> dispItems = new ArrayList<>();
-        for (Item item : lookup.getData()) {
-            dispItems.add(dispItemFromItem(item, accountKey));
-        }
-        return new PagedData<>(
-                lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(), dispItems
-        );
+        PagedData<Item> lookup = childForAssetCatalog(assetCatalogKey, pagingInfo);
+        return toDispPagedData(lookup, accountKey);
     }
 
-    @SuppressWarnings("DuplicatedCode")
     @Override
     public PagedData<DispItem> childForAssetCatalogRootDisp(
             StringIdKey accountKey, LongIdKey assetCatalogKey, PagingInfo pagingInfo
     ) throws ServiceException {
-        PagedData<Item> lookup = itemMaintainService.lookup(
-                ItemMaintainService.CHILD_FOR_ASSET_CATALOG_ROOT, new Object[]{assetCatalogKey}, pagingInfo
-        );
-        List<DispItem> dispItems = new ArrayList<>();
-        for (Item item : lookup.getData()) {
-            dispItems.add(dispItemFromItem(item, accountKey));
-        }
-        return new PagedData<>(
-                lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(), dispItems
-        );
+        PagedData<Item> lookup = childForAssetCatalogRoot(assetCatalogKey, pagingInfo);
+        return toDispPagedData(lookup, accountKey);
     }
 
-    @SuppressWarnings("DuplicatedCode")
     @Override
     public PagedData<DispItem> childForParentDisp(
             StringIdKey accountKey, LongIdKey parentKey, PagingInfo pagingInfo
     ) throws ServiceException {
-        PagedData<Item> lookup = itemMaintainService.lookup(
-                ItemMaintainService.CHILD_FOR_PARENT, new Object[]{parentKey}, pagingInfo
-        );
-        List<DispItem> dispItems = new ArrayList<>();
-        for (Item item : lookup.getData()) {
-            dispItems.add(dispItemFromItem(item, accountKey));
-        }
-        return new PagedData<>(
-                lookup.getCurrentPage(), lookup.getTotalPages(), lookup.getRows(), lookup.getCount(), dispItems
-        );
+        PagedData<Item> lookup = childForParent(parentKey, pagingInfo);
+        return toDispPagedData(lookup, accountKey);
     }
 
-    private DispItem dispItemFromItem(Item item, StringIdKey inspectAccountKey)
+    @Override
+    public PagedData<DispItem> childForAssetCatalogNameLikeDisp(
+            StringIdKey accountKey, LongIdKey assetCatalogKey, String pattern, PagingInfo pagingInfo
+    ) throws ServiceException {
+        PagedData<Item> lookup = childForAssetCatalogNameLike(assetCatalogKey, pattern, pagingInfo);
+        return toDispPagedData(lookup, accountKey);
+    }
+
+    private DispItem toDisp(Item item, StringIdKey inspectAccountKey)
             throws ServiceException {
         DispAssetCatalog assetCatalog = null;
         if (Objects.nonNull(item.getAssetCatalogKey())) {
@@ -181,6 +167,18 @@ public class ItemResponseServiceImpl implements ItemResponseService {
         return DispItem.of(item, assetCatalog, typeIndicator, itemLabels, hasNoChild);
     }
 
+    private PagedData<DispItem> toDispPagedData(PagedData<Item> itemPagedData, StringIdKey inspectAccountKey)
+            throws ServiceException {
+        List<DispItem> dispItems = new ArrayList<>();
+        for (Item item : itemPagedData.getData()) {
+            dispItems.add(toDisp(item, inspectAccountKey));
+        }
+        return new PagedData<>(
+                itemPagedData.getCurrentPage(), itemPagedData.getTotalPages(), itemPagedData.getRows(),
+                itemPagedData.getCount(), dispItems
+        );
+    }
+
     @Override
     public LongIdKey createItem(StringIdKey userKey, ItemCreateInfo itemCreateInfo) throws
             ServiceException {
@@ -195,5 +193,50 @@ public class ItemResponseServiceImpl implements ItemResponseService {
     @Override
     public void removeItem(StringIdKey userKey, LongIdKey itemKey) throws ServiceException {
         itemOperateService.removeItem(userKey, itemKey);
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    public List<LongIdKey> pathFromRoot(LongIdKey key) throws ServiceException {
+        // 获取当前的权限组。
+        Item item = itemMaintainService.get(key);
+
+        // 如果权限组没有父节点，则返回空列表。
+        if (Objects.isNull(item.getParentKey())) {
+            return new ArrayList<>();
+        }
+
+        // 定义结果列表。
+        List<LongIdKey> result = new ArrayList<>();
+
+        // 获取权限组的父节点主键，作为当前节点主键。
+        LongIdKey anchorKey = item.getParentKey();
+
+        // 将当前节点主键添加到结果列表中。
+        result.add(anchorKey);
+
+        // 循环获取父节点的父节点，直到父节点为空。
+        while (Objects.nonNull(anchorKey)) {
+            // 获取当前节点。
+            item = itemMaintainService.get(anchorKey);
+
+            // 获取当前节点的父节点主键。
+            LongIdKey parentKey = item.getParentKey();
+
+            // 如果当前节点没有父节点，则跳出循环。
+            if (Objects.isNull(parentKey)) {
+                break;
+            }
+
+            // 将父节点主键添加到结果列表中。
+            result.add(parentKey);
+
+            // 将父节点主键作为当前节点主键。
+            anchorKey = parentKey;
+        }
+
+        // 将结果列表反转，并返回。
+        Collections.reverse(result);
+        return result;
     }
 }