+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.system.PermissionGroupController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PbItemController。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.PbNodeController。
+  - com.dwarfeng.familyhelper.webapi.node.controller.v1.assets.ItemController。
 
 - 新建 Controller。
   - com.dwarfeng.familyhelper.webapi.node.controller.v1.life.ActivityTemplateDriverSupportController。